backend/requirements.txt
google-genai>=1.0.0
ddgs>=6.0.0
pydantic>=2.0.0
modal>=0.60.0
